design ui gnn model/requirements.txt
fastapi
uvicorn
torch
torch-geometric
rdkit
biopython
numpy
